check ebpf kernel
