+++ src/cli/src/utils.rs
@@ -1,5 +1,6 @@
 use anyhow::Context;
 use std::path::Path;
+use std::process::Command;
 
 pub fn ensure_file_can_be_created<P: AsRef<Path>>(file_path: P) -> anyhow::Result<()> {
     let file_path = file_path.as_ref();
@@ -25,4 +26,25 @@ pub fn check_sudo_privileges() {
         let sudo_command = format!("sudo {} {}", current_exe.display(), args[1..].join(" "));
         println!("Try runing with elevated privileges:\n {}", sudo_command);
     }
+}
+
+pub fn get_kernel_version() -> Option<(u32, u32)> {
+    let kernel_version = Command::new("uname")
+        .arg("-r")
+        .output()
+        .ok()
+        .and_then(|output| {
+            String::from_utf8(output.stdout).ok().and_then(|version| {
+                let parts: Vec<&str> = version.trim().split('.').collect();
+                if parts.len() >= 2 {
+                    let major = parts[0].parse::<u32>().ok()?;
+                    let minor = parts[1].parse::<u32>().ok()?;
+                    Some((major, minor))
+                } else {
+                    None
+                }
+            })
+        });
+
+    kernel_version
 }
